IIFE / forEach
I protected the base Pokemon within an IIFE and then made my code to print to the page use forEach instead on i++. Then using keys to access the data within the IIFE I made the document write to the page. I also added a Pokemon using the keys as well.